requirements.txt
anyio==4.10.0
APScheduler==3.10.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
paho-mqtt==1.6.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-telegram-bot==20.8
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
